Add Windows e2e integration tests, SP5/IAM integration, and clippy fixes
- New windows-e2e-tests crate with 17 cross-process integration tests
  covering named pipes, credential extraction, handle duplication, and
  CAL control channels
- Rust-based test coordinator replacing Unix expect scripts for Windows
- SP5 IAM integration with service builders and CAL control channel
- Fix clippy warnings across PAL, CAL, and BB crates (unused imports,
  dead code, wrong_self_convention, manual_div_ceil, clone_on_copy,
  disallowed_types, unnecessary_cast, needless_range_loop)
- cargo fmt --all clean, cargo clippy clean on all modified crates

+++ iceoryx2-cal/src/security/credentials.rs
@@ -115,7 +115,7 @@ impl ProcessCredentials {
         use iceoryx2_bb_posix::user::User;
 
         let process = Process::from_self();
-        let pid = process.id().value() as u32;
+        let pid = process.id().value();
 
         // Get uid and gid from user - fallback to 0 if unavailable
         let (uid, gid) = match User::from_self() {
@@ -143,7 +143,7 @@ impl ProcessCredentials {
         use iceoryx2_bb_posix::user::User;
 
         let process = Process::from_self();
-        let pid = process.id().value() as u32;
+        let pid = process.id().value();
 
         // Get uid and gid from user - fallback to 0 if unavailable
         let (uid, gid) = match User::from_self() {
@@ -172,7 +172,7 @@ impl ProcessCredentials {
         use iceoryx2_bb_posix::user::User;
 
         let process = Process::from_self();
-        let pid = process.id().value() as u32;
+        let pid = process.id().value();
 
         // Get uid and gid from user - fallback to 0 if unavailable
         let (uid, gid) = match User::from_self() {
@@ -194,7 +194,7 @@ impl ProcessCredentials {
         use iceoryx2_bb_posix::user::User;
 
         let process = Process::from_self();
-        let pid = process.id().value() as u32;
+        let pid = process.id().value();
 
         // Get uid and gid from user - fallback to 0 if unavailable
         let (uid, gid) = match User::from_self() {
